requirements.txt
# Web Framework
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1

# Database
pymysql==1.1.1
cryptography==42.0.8
sqlalchemy==2.0.31

# Data & Analysis
pandas==2.2.2
pandas-ta==0.3.14b
numpy==1.26.4
requests==2.32.3

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.1

# Utilities
pytz==2024.1
